+++ utils/constants.js
@@ -184,3 +184,54 @@ export const CSNAV_ITEMS = {
   
 
 }
+
+export const NAV_ELEMENTS_VARIANT = {
+  bank: {
+    navLinks: [
+      {
+        text: "Summary",
+        href: "/bank",
+      },
+      { text: "Transfers", href: "/bank" },
+      { text: "Deposits", href: "/bank" },
+      { text: "External Accounts", href: "/bank" },
+      { text: "Statements", href: "/bank" },
+    ],
+    navLinkColor: "gradient-bank",
+    popoverMessage: "Thank you for banking with us, ",
+    logoImg: COMPANY_LOGOS["bank"].horizontal,
+  },
+  investment: {
+    navLinks: [
+      { text: "Accounts & Trade", href: "/investment" },
+      { text: "Planning", href: "/investment" },
+      { text: "News", href: "/investment" },
+      { text: "Investment Products", href: "/investment" },
+      { text: "About Us", href: "/investment" },
+    ],
+    navLinkColor: "gradient-investment",
+    popoverMessage: "Thank you for investing with us, ",
+    logoImg: COMPANY_LOGOS["investment"].horizontal,
+  },
+  market: {
+    navLinks: [
+      { text: "All", href: "/marketplace" },
+      { text: "Account", href: "/marketplace" },
+      { text: "Buy Again", href: "/marketplace" },
+      { text: "Today's Deals", href: "/marketplace" },
+      { text: "Sale", href: "/marketplace" },
+    ],
+    navLinkColor: "gradient-experimentation",
+    popoverMessage: "Thank you for shopping with us, ",
+    logoImg: COMPANY_LOGOS["market"].horizontal,
+  },
+  airlines: {
+    navLinks: [
+      { text: "Book", href: "/airways" },
+      { text: "Check-In", href: "/airways" },
+    ],
+    navLinkColor: "gradient-airline-buttons",
+    popoverMessage: "Thank you for flying with us, ",
+    logoImg: COMPANY_LOGOS["airlines"].horizontal,
+  },
+};
